+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Fixed bug with no checking operation error on `discovery.Client` calls
 * Allowed zero create session timeout in `ydb.WithSessionPoolCreateSessionTimeout(timeout)` (less than or equal to zero - no used timeout on create session request)
 * Added examples with own `go.mod`
+* Supported `scheme.EntryTopic` path child entry in `sugar.RemoveRecursive`
 * Renamed private `ydb.connection` type to public `ydb.Driver` type
 * Marked as deprecated `ydb.Connection` interface
 * Changed result type of `ydb.Open` from `ydb.Connection` interface to `*ydb.Driver`